+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add `gulp sentry:release`
 - Add ES2015 polyfills using `babel-polyfill`
 - Hide statistics for minions with HIDE\_STATS (#128, @jleclanche)
+- Add `Launcher.enableKeybindings()` and `Launcher.disableKeybindings()`
 
 ### Changed
 - Replace own HearthstoneJSON implementation with common one